+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- added: RLUSD on Ethereum and Ripple networks
+- added: SOLO on Ripple network
+
 ## 4.37.3 (2025-02-28)
 
 - changed: (Abstract) Ignore internal transactions